chore: file structure reorganization
