+++ packages/client/src/events/participant.ts
@@ -1,5 +1,5 @@
-import { Dispatcher } from '../rtc/Dispatcher';
-import { VisibilityState } from '../rtc/types';
+import { Dispatcher } from '../rtc';
+import { VisibilityState } from '../types';
 import { CallState } from '../store';
 
 /**
